day17; needed hints but mostly fought with myself
my min heap implementation had a silly bug where I forgot to move the last item, move it to the top and then heapify that down when popping :facepalm: I was just heapifying down the new 0th item which of course isn't what it needed.

anyway, this works, but is surprisingly slow. I didn't profile it but I'm guessing its my heap impl.