2024 Day 15
I like these ones. had a nasty bug in part 2 where I didn't recursively check both sides of boxes for movement and it worked for the test input :sweat:. Had to add a check for invalid boxes and then was able to write a test case for it to debug. :ok: